+++ src/mesido/esdl/asset_to_component_base.py
@@ -295,6 +295,11 @@ class _AssetToComponentBase:
             "fixedMaintenanceCosts": "optional",
             "fixedOperationalCosts": "optional",
         },
+        "electricity_import": {
+            "investmentCosts": "optional",
+            "installationCosts": "optional",
+            "variableOperationalCosts": "optional",
+        },
         "heat_source": {  # Includes GeothermalSource, ResidualHeatSource, HeatProducer,
             # GasHeater, ElectricBoiler
             "investmentCosts": "required",
@@ -349,6 +354,8 @@ class _AssetToComponentBase:
         "CoolingDemand": "heat_demand",
         "Electrolyzer": "electrolyzer",
         "ElectricBoiler": "heat_source",
+        "Import": "electricity_import",
+        "ElectricityProducer": "electricity_source",
         "GasDemand": "gas_demand",
         "GasHeater": "heat_source",
         "GasStorage": "gas_tank_storage",