+++ src/mesido/financial_mixin.py
@@ -1163,9 +1163,15 @@ class FinancialMixin(
             variable_operational_cost_coefficient = parameters[  # euro / Wh
                 f"{es}.variable_operational_cost_coefficient"
             ]
+
+            price_profile = self.__get_electricity_price_profile_or_zero()
+
             sum_ = ca.sum1(
                 variable_operational_cost_coefficient * elec_produced_w[1:] * timesteps_hr
             )  # [euro/Wh] * [W] * [hr]
+            if es in self.energy_system_components.get("electricity_import", []):
+                sum_ += ca.sum1(price_profile[1:] * elec_produced_w[1:] * timesteps_hr)
+
             constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         # for a in self.heat_network_components.get("ates", []):