+++ tests/utils_tests.py
@@ -12,8 +12,6 @@ from mesido.head_loss_class import HeadLossOption
 
 import numpy as np
 
-from rtctools.optimization.timeseries import Timeseries
-
 
 def __get_out_port_temp_profile(solution, asset_name, asset_type):
     """
@@ -218,7 +216,7 @@ def heat_to_discharge_test(solution, results, atol=1e-2, rtol=1.0e-4):
         if temp_profile is not None:
             supply_t = temp_profile.values
             supply_temp_profiles.append(temp_profile.values)
-        print(d, max(abs(results[f"{d}.HeatOut.Heat"] - results[f"{d}.Q"] * rho * cp * supply_t)))
+        # print(d, max(abs(results[f"{d}.HeatOut.Heat"] - results[f"{d}.Q"] * rho * cp * supply_t)))
 
         np.testing.assert_allclose(
             results[f"{d}.HeatOut.Heat"],
@@ -741,7 +739,6 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
     assets = [
         *solution.energy_system_components.get("heat_source", []),
         *solution.energy_system_components.get("ates", []),
-        *solution.energy_system_components.get("low_temperature_ates", []),
         *solution.energy_system_components.get("heat_pump", []),
         *solution.energy_system_components.get("pump", []),
         *solution.energy_system_components.get("heat_exchanger", []),
@@ -749,6 +746,7 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
         *solution.energy_system_components.get("airco", []),
         *solution.energy_system_components.get("heat_demand", []),
         *solution.energy_system_components.get("cold_demand", []),
+        *solution.energy_system_components.get("electricity_source", []),
         *transport_assets,
     ]
 
@@ -759,17 +757,20 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
         for name, edr_class_name in _AssetToComponentBase.STEEL_S1_PIPE_EDR_ASSETS.items()
     ]
 
+    assets_may_have_electricity_price_profile = [
+        *solution.energy_system_components.get("heat_source_elec", []),
+        *solution.energy_system_components.get("air_water_heat_pump", []),
+        *[
+            hp_asset
+            for hp_asset in solution.energy_system_components.get("heat_pump", [])
+            if hp_asset not in solution.energy_system_components.get("heat_pump_elec", [])
+        ],
+        *solution.energy_system_components.get("electricity_import", []),
+    ]
+
     edr_pipes = json.load(open(Path(__file__).parent.parent / "src/mesido/esdl/_edr_pipes.json"))
 
-    if len(solution.get_electricity_carriers().keys()) == 1:
-        try:
-            price_profile = solution.get_timeseries(
-                f"{list(solution.get_electricity_carriers().values())[0]['name']}.price_profile"
-            )
-        except KeyError:
-            price_profile = Timeseries(solution.times(), np.zeros(len(solution.times())))
-    else:
-        price_profile = Timeseries(solution.times(), np.zeros(len(solution.times())))
+    price_profile = solution._FinancialMixin__get_electricity_price_profile_or_zero()
 
     total_investment_cost = 0.0
     total_installation_cost = 0.0
@@ -785,7 +786,10 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
         ):
             pass
         elif costs_esdl_asset is None:
-            continue
+            if asset in assets_may_have_electricity_price_profile:
+                pass  # Assets include e-price in var-opex, but no esdl cost attributes are defined
+            elif asset not in assets_may_have_electricity_price_profile:
+                continue
 
         # Investment Cost
         investment_cost = 0.0
@@ -798,8 +802,10 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
                     edr_pipes, parameters[f"{asset}.diameter"]
                 )
             else:
-                investment_cost_info = _get_esdl_scaled_cost(
-                    costs_esdl_asset.investmentCosts, esdl.UnitEnum.WATT
+                investment_cost_info = (
+                    _get_esdl_scaled_cost(costs_esdl_asset.investmentCosts, esdl.UnitEnum.WATT)
+                    if costs_esdl_asset is not None
+                    else 0.0
                 )
                 if asset in solution.energy_system_components.get("heat_buffer", []):
                     np.testing.assert_allclose(
@@ -847,8 +853,8 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
         else:
             if results[f"{asset}__max_size"] > 1e-8:
                 installation_cost = (
-                    costs_esdl_asset.installationCosts.value
-                    if costs_esdl_asset.installationCosts is not None
+                    _get_esdl_scaled_cost(costs_esdl_asset.installationCosts, esdl.UnitEnum.WATT)
+                    if costs_esdl_asset is not None
                     else 0.0
                 )
             np.testing.assert_allclose(installation_cost, results[f"{asset}__installation_cost"])
@@ -859,11 +865,15 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
         if asset in transport_assets:
             np.testing.assert_allclose(0.0, results[f"{asset}__fixed_operational_cost"])
         else:
-            fix_op_costs_esdl = _get_esdl_scaled_cost(
-                costs_esdl_asset.fixedOperationalCosts, esdl.UnitEnum.WATT
+            fix_op_costs_esdl = (
+                _get_esdl_scaled_cost(costs_esdl_asset.fixedOperationalCosts, esdl.UnitEnum.WATT)
+                if costs_esdl_asset is not None
+                else 0.0
             )
-            fix_maint_costs_esdl = _get_esdl_scaled_cost(
-                costs_esdl_asset.fixedMaintenanceCosts, esdl.UnitEnum.WATT
+            fix_maint_costs_esdl = (
+                _get_esdl_scaled_cost(costs_esdl_asset.fixedMaintenanceCosts, esdl.UnitEnum.WATT)
+                if costs_esdl_asset is not None
+                else 0.0
             )
             np.testing.assert_allclose(
                 parameters[f"{asset}.fixed_operational_cost_coefficient"],
@@ -884,8 +894,12 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
         if asset in transport_assets:
             np.testing.assert_allclose(0.0, results[f"{asset}__variable_operational_cost"])
         else:
-            var_op_costs_esdl = _get_esdl_scaled_cost(
-                costs_esdl_asset.variableOperationalCosts, esdl.UnitEnum.WATTHOUR
+            var_op_costs_esdl = (
+                _get_esdl_scaled_cost(
+                    costs_esdl_asset.variableOperationalCosts, esdl.UnitEnum.WATTHOUR
+                )
+                if costs_esdl_asset is not None
+                else 0.0
             )
             np.testing.assert_allclose(
                 parameters[f"{asset}.variable_operational_cost_coefficient"], var_op_costs_esdl
@@ -899,7 +913,7 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
                 if parameters[f"{asset}.include_head_loss_variables"]:
                     pump_power = results[f"{asset}.Pump_power"]
                     eff = parameters[f"{asset}.pump_efficiency"]
-                    pump_cost = sum(price_profile.values[1:] * pump_power[1:] * timesteps_hr / eff)
+                    pump_cost = sum(price_profile[1:] * pump_power[1:] * timesteps_hr / eff)
                 else:
                     pump_cost = 0.0
 
@@ -913,7 +927,7 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
                 if parameters[f"{asset}.include_head_loss_variables"]:
                     pump_power = results[f"{asset}.Pump_power"]
                     eff = parameters[f"{asset}.pump_efficiency"]
-                    pump_cost = sum(price_profile.values[1:] * pump_power[1:] * timesteps_hr / eff)
+                    pump_cost = sum(price_profile[1:] * pump_power[1:] * timesteps_hr / eff)
                 else:
                     pump_cost = 0.0
 
@@ -947,6 +961,8 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
                 nominator_vector = results[f"{asset}.Heat_demand"]
             elif asset in solution.energy_system_components.get("cold_demand", []):
                 nominator_vector = results[f"{asset}.Cold_demand"]
+            elif asset in solution.energy_system_components.get("electricity_source", []):
+                nominator_vector = results[f"{asset}.Electricity_source"]  # [W]
             else:
                 raise AssertionError(
                     f"Asset '{esdl_asset.name}' is not handled in the variable operational"
@@ -959,13 +975,11 @@ def cost_calculation_test(solution, results, check_objective_function=False, ato
 
             variable_operational_cost += pump_cost
 
-            if (len(solution.get_electricity_carriers().keys()) > 0) and asset in [
-                *solution.energy_system_components.get("heat_source_elec", []),
-                *solution.energy_system_components.get("air_water_heat_pump", []),
-                *solution.energy_system_components.get("heat_pump", []),
-            ]:
+            if (
+                len(solution.get_electricity_carriers().keys()) > 0
+            ) and asset in assets_may_have_electricity_price_profile:
                 variable_operational_cost += sum(
-                    price_profile.values[1:] * nominator_vector[1:] * timesteps_hr / denominator
+                    price_profile[1:] * nominator_vector[1:] * timesteps_hr / denominator
                 )
 
             np.testing.assert_allclose(